feat:add addLike, deleteLike, top films in class Film

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -9,8 +9,8 @@ public interface FilmStorage {
     void add(Film film);
     void update(Film film);
     void delete();
-    boolean isExist(int filmId);
+    boolean isExist(long filmId);
     ArrayList<Film> getAll();
-    Film getById(int id);
+    Film getById(long id);
 
 }
